paginated range query (#135)
This commit adds a range query API to the data query executor.1

+++ pkg/bcdb/ledger.go
@@ -101,7 +101,7 @@ func (l *ledger) GetLedgerPath(startBlock, endBlock uint64) (*LedgerPath, error)
 }
 
 func (l *ledger) GetTransactionProof(blockNum uint64, txIndex int) (*TxProof, error) {
-	path := constants.URLTxProof(blockNum, txIndex)
+	path := constants.URLTxProof(blockNum, uint64(txIndex))
 	resEnv := &types.GetTxProofResponseEnvelope{}
 	err := l.handleRequest(
 		path,
